pyproject.toml
[project]
name = "hitshack"
version = "0.1.0"
description = ""
authors = [
    {name = "Trebias",email = "unitsss11@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "django (>=5.2.3,<6.0.0)",
    "python-telegram-bot (>=22.1,<23.0)",
    "requests (>=2.32.4,<3.0.0)",
    "scikit-learn (>=1.7.0,<2.0.0)",
    "opencage (>=3.2.0,<4.0.0)",
    "google-api-python-client (>=2.172.0,<3.0.0)",
    "google-auth (>=2.40.3,<3.0.0)",
    "google-auth-oauthlib (>=1.2.2,<2.0.0)",
    "google-auth-httplib2 (>=0.2.0,<0.3.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
